// proto/bike/BikeResponse.ts
// Original file: proto/bike.proto


export interface BikeResponse {
  'id'?: (number);
  'brand'?: (string);
  'model'?: (string);
  'color'?: (string);
  'year'?: (number);
}

export interface BikeResponse__Output {
  'id'?: (number);
  'brand'?: (string);
  'model'?: (string);
  'color'?: (string);
  'year'?: (number);
}